requirements.txt
# Core data manipulation and analysis
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
python-dotenv>=0.19.0

# Financial data and APIs
yfinance>=0.2.0
fredapi>=0.4.3
quandl>=3.7.0
alpha-vantage>=2.3.1

# Statistical analysis and machine learning
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Risk management and portfolio optimization
cvxpy>=1.3.0
pyportfolioopt>=1.5.0

# Database and data storage
sqlalchemy>=2.0.0
h5py>=3.8.0

# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.7.0
flake8>=6.0.0

# Timezone handling
pytz>=2023.3

# Excel file handling
openpyxl>=3.1.0
xlrd>=2.0.1